@@ -1,6 +1,6 @@
 {
   "name": "NativeSigner",
-  "version": "3.0.0-beta",
+  "version": "3.0.5-beta",
   "private": true,
   "license": "GPL-3.0",
   "engines": {
@@ -8,7 +8,8 @@
     "yarn": "^1.3.2"
   },
   "scripts": {
-    "android": "yarn run build-rust-android && react-native run-android --appIdSuffix \"dev\"",
+    "android": "yarn run build-rust-android && react-native run-android",
+    "build-prod:ios": "NODE_OPTIONS=--max_old_space_size=8192 react-native bundle --dev false --entry-file index.js --bundle-output ios/main.jsbundle --platform ios",
     "build-rust-ios": "cd rust/signer && make ios",
     "build-rust-android": "cd rust/signer && make android",
     "clean": "watchman watch-del-all && rm -rf /tmp/metro-bundler-cache-* && rm -rf /tmp/haste-map-react-native-packager-* && rm -rf node_modules/ && yarn cache clean --force",
@@ -20,32 +21,32 @@
     "test": "jest"
   },
   "dependencies": {
-    "@plugnet/util": "^0.94.100",
+    "@plugnet/util": "^1.1.100",
     "@plugnet/wasm-crypto-js": "^0.11.102",
-    "@polkadot/api": "^0.90.0-beta.66",
-    "@polkadot/reactnative-identicon": "^0.42.0-beta.26",
-    "@polkadot/util": "^1.0.1",
-    "@polkadot/util-crypto": "1.1.0-beta.1",
+    "@polkadot/api": "^0.91.0-beta.10",
+    "@polkadot/reactnative-identicon": "^0.42.1",
+    "@polkadot/util": "^1.1.1",
+    "@polkadot/util-crypto": "1.1.1",
     "@polkadot/wasm-crypto": "^0.13.1",
-    "@react-native-community/netinfo": "^4.1.1",
+    "@react-native-community/netinfo": "^4.1.5",
     "@tradle/react-native-http": "^2.0.1",
     "bignumber.js": "^9.0.0",
     "hoist-non-react-statics": "^3.3.0",
     "node-libs-react-native": "^1.0.3",
     "prop-types": "^15.6.1",
     "react": "^16.9.0",
-    "react-native": "0.60.4",
-    "react-native-camera": "^3.0.0",
-    "react-native-gesture-handler": "^1.3.0",
-    "react-native-keyboard-aware-scroll-view": "^0.8.0",
+    "react-native": "0.60.5",
+    "react-native-camera": "^3.3.0",
+    "react-native-gesture-handler": "^1.4.1",
+    "react-native-keyboard-aware-scroll-view": "^0.9.1",
     "react-native-markdown-renderer": "^3.2.8",
     "react-native-popup-menu": "^0.15.6",
     "react-native-randombytes": "^3.5.3",
     "react-native-secure-storage": "git+https://github.com/paritytech/react-native-secure-storage.git",
-    "react-native-svg": "^9.5.3",
+    "react-native-svg": "^9.7.1",
     "react-native-tabs": "^1.0.9",
     "react-native-vector-icons": "^6.6.0",
-    "react-navigation": "^3.11.1",
+    "react-navigation": "^3.12.1",
     "readable-stream": "^3.4.0",
     "unstated": "^2.1.1",
     "vm-browserify": "1.1.0"
@@ -57,17 +58,17 @@
     "@babel/plugin-transform-runtime": "^7.5.0",
     "@babel/runtime": "^7.5.4",
     "@react-native-community/eslint-config": "^0.0.5",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.8.0",
+    "babel-eslint": "^10.0.3",
+    "babel-jest": "^24.9.0",
     "babel-plugin-rewrite-require": "^1.14.5",
-    "eslint": "^6.0.1",
-    "jest": "^24.8.0",
-    "jetifier": "^1.6.3",
-    "metro-react-native-babel-preset": "^0.55.0",
+    "eslint": "^6.2.2",
+    "jest": "^24.9.0",
+    "jetifier": "^1.6.4",
+    "metro-react-native-babel-preset": "^0.56.0",
     "pre-git": "^3.14.0",
     "prettier": "1.18.2",
-    "react-test-renderer": "16.8.6",
-    "reactotron-react-native": "^3.6.4"
+    "react-test-renderer": "16.9.0",
+    "reactotron-react-native": "^3.6.5"
   },
   "release": {
     "analyzeCommits": "simple-commit-message"
