@@ -4,7 +4,7 @@ export default `
 ## LEGAL WARNING SHORT VERSION
 ### Disclaimer of Liability and Warranties
 - The user expressly acknowledges and agrees that Parity Technologies Limited makes the Parity Signer application available to the user at the user's sole risk.
-- The user represents that the user has an adequate understanding of the risks, usage and intricacies of public and private key cryptography, cryptographic tokens, mobile wallet technology, blockchain-based open source software, the Ethereum platform and Ether (ETH).
+- The user represents that the user has an adequate understanding of the risks, usage and intricacies of public and private key cryptography, cryptographic tokens, mobile wallet technology, blockchain-based open source software, the various platforms that Signer supports, including the Ethereum platform and any platforms built on Substrate including Kusama and Polkadot, and the blockchain tokens native to the various platforms supported by Signer, including Ether (ETH), Kusamas (KSM) and DOTs.
 - The user acknowledges and agrees that, to the fullest extent permitted by any applicable law, the disclaimers of liability contained herein apply to any and all damages or injury whatsoever caused by or related to risks of, use of, or inability to use, the Parity Signer application under any cause or action whatsoever of any kind in any jurisdiction, including, without limitation, actions for breach of warranty, breach of contract or tort (including negligence) and that Parity Technologies Limited shall not be liable for any indirect, incidental, special, exemplary or consequential damages, including for loss of profits, goodwill or data.
 - Some jurisdictions do not allow the exclusion of certain warranties or the limitation or exclusion of liability for certain types of damages. Therefore, some of the above limitations in this section may not apply to a user. In particular, nothing in these terms shall affect the statutory rights of any user or limit or exclude liability for death or physical injury arising from the negligence or wilful misconduct of Parity Technologies Limited or for fraud or fraudulent misrepresentation.
 - All rights reserved by Parity Technologies Limited. Licensed to the public under the GNU General Public License Version 3: https://opensource.org/licenses/GPL-3.0
@@ -16,13 +16,13 @@ The following Terms and Conditions ("Terms") govern the use of Parity Technologi
 The User acknowledges the following serious risks to any users of Parity Signer and expressly agrees not to hold liable Parity Technologies or the Parity Technologies Team should any of these risks occur:
 
 ### Risk of Security Weaknesses in the Parity Core Infrastructure Software
-Parity Signer uses open-source libraries and components developed by third parties. While Parity Technologies Limited generally aims to use only widely adopted open-source technology and develop it in line with industry standards, such open-source technology may contain bugs and errors and may not function correctly in all circumstances. As a result, there is a risk that Parity Technologies or the Parity Technologies Team may have introduced unintentional weaknesses or bugs into the core infrastructural elements of Parity Signer causing the loss of private keys stored in one or more user accounts in the application, Ethereum tokens ("ETH") or sums of other valued tokens.
+Parity Signer uses open-source libraries and components developed by third parties. While Parity Technologies Limited generally aims to use only widely adopted open-source technology and develop it in line with industry standards, such open-source technology may contain bugs and errors and may not function correctly in all circumstances. As a result, there is a risk that Parity Technologies or the Parity Technologies Team may have introduced unintentional weaknesses or bugs into the core infrastructural elements of Parity Signer causing the loss of private keys stored in one or more user accounts in the application and sums of valued blockchain tokens such as ETH, KSMs or DOTs.
 
 ### Risk of Weaknesses or Exploitable Breakthroughs in the Field of Cryptography
-Cryptography is an art, not a science, and the state of the art can advance over time. Advances in code cracking, or technical advances such as the development of quantum computers, could present risks to cryptocurrencies and Parity Signer, which could result in the theft or loss of private keys stored in one or more user accounts in the Parity Signer application, ETH or sums of other valued tokens. To the extent possible, Parity Technologies intends to update the software underlying Parity Signer to account for any advances in cryptography and to incorporate additional security measures, but it cannot predict the future of cryptography or guarantee that any security updates will be made, timely or successful.
+Cryptography is an art, not a science, and the state of the art can advance over time. Advances in code cracking, or technical advances such as the development of quantum computers, could present risks to cryptocurrencies and Parity Signer, which could result in the theft or loss of private keys stored in one or more user accounts in the Parity Signer application, ETH, KSMs, DOTs or sums of other valued tokens. To the extent possible, Parity Technologies intends to update the software underlying Parity Signer to account for any advances in cryptography and to incorporate additional security measures, but it cannot predict the future of cryptography or guarantee that any security updates will be made, timely or successful.
 
 ### Risk of Mining Attacks
-The blockchains for which Parity Signer creates accounts are susceptible to mining attacks, including but not limited to double-spend attacks, majority mining power attacks, "selfish-mining" attacks, and race condition attacks. Any successful attacks present a risk to the ecosystems of those blockchains, for example in respect of the Ethereum ecosystem a successful attack would present risks to the expected proper execution and sequencing of ETH transactions, and the expected proper execution and sequencing of contract computations. Despite the efforts of Parity Technologies and the Parity Technologies Team, known or novel mining attacks may be successful.
+Some of the blockchains for which Parity Signer creates accounts are susceptible to mining attacks, including but not limited to double-spend attacks, majority mining power attacks, "selfish-mining" attacks, and race condition attacks. Any successful attacks present a risk to the ecosystems of those blockchains, for example in respect of the Ethereum ecosystem a successful attack would present risks to the expected proper execution and sequencing of ETH transactions, and the expected proper execution and sequencing of contract computations. Despite the efforts of Parity Technologies and the Parity Technologies Team, known or novel mining attacks may be successful.
 
 ### Risk of Rapid Adoption and Insufficiency of Computational Application Processing Power on the Ethereum Network
 If Ethereum is rapidly adopted, the demand for transaction processing and distributed application computations could rise dramatically and at a pace that exceeds the rate with which ETH miners can bring online additional mining power. Under such a scenario, the entire Ethereum ecosystem could become destabilized, due to the increased cost of running distributed applications. In turn, this could dampen interest in the Ethereum ecosystem and ETH. Insufficiency of computational resources and an associated rise in the price of ETH could result in businesses being unable to acquire scarce computational resources to run their distributed applications. This would represent revenue losses to businesses or worst case, cause businesses to cease operations because such operations have become uneconomical due to distortions in the crypto-economy.
