@@ -16,33 +16,67 @@
 
 // @flow
 
+import extrinsicsFromMeta from '@polkadot/api-metadata/extrinsics/fromMetadata';
+import { GenericCall, Metadata } from '@polkadot/types';
+import Call from '@polkadot/types/primitive/Generic/Call';
+
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { StyleSheet, Text, View, ViewPropTypes } from 'react-native';
 
-import colors from '../colors';
 import fonts from '../fonts';
+import colors from '../colors';
+import { SUBSTRATE_NETWORK_LIST, SubstrateNetworkKeys } from '../constants';
+import kusamaMetadata from '../util/static-kusama';
+import substrateDevMetadata from '../util/static-substrate';
 
 export default class PayloadDetailsCard extends React.PureComponent {
   static propTypes = {
     description: PropTypes.string.isRequired,
     payload: PropTypes.object,
+    prefix: PropTypes.number.isRequired,
     signature: PropTypes.string,
     style: ViewPropTypes.style
   };
 
-  render() {
-    const { description, payload, signature, style } = this.props;
+  state = {
+    fallback: false
+  };
+
+  constructor(props) {
+    super(props);
+
+    let metadata;
+    if (this.props.prefix === SUBSTRATE_NETWORK_LIST[SubstrateNetworkKeys.KUSAMA].prefix) {
+      metadata = new Metadata(kusamaMetadata);
+    } else if (this.props.prefix === SUBSTRATE_NETWORK_LIST[SubstrateNetworkKeys.SUBSTRATE_DEV].prefix) {
+      metadata = new Metadata(substrateDevMetadata);
+    } 
     
+    if (!metadata) {
+      this.setState({
+        fallback: true
+      });
+    }
+
+    const extrinsics = extrinsicsFromMeta(metadata);
+    GenericCall.injectMethods(extrinsics);
+  }
+
+
+  render() {
+    const { fallback } = this.state;
+    const { description, payload, signature, style } = this.props;
+
     return (
       <View style={[styles.body, style]}>
         <Text style={styles.titleText}>{description}</Text>
         {
-          payload && (
+          !!payload && (
             <View style={{ padding: 5, paddingVertical: 2 }}>
               <ExtrinsicPart label='Block Hash' value={payload.blockHash.toString()} />
-              <ExtrinsicPart label='Method' value={payload.method.toString()} />
-              <ExtrinsicPart label='Era' value={payload.era.toString()} />
+              <ExtrinsicPart label='Method' value={fallback ? payload.method.toString() : payload.method} />
+              <ExtrinsicPart label='Era' value={fallback ? payload.era.toString() : payload.era} />
               <ExtrinsicPart label='Nonce' value={payload.nonce.toString()} />
               <ExtrinsicPart label='Tip' value={payload.tip.toString()} />
               <ExtrinsicPart label='Genesis Hash' value={payload.genesisHash.toString()} />
@@ -50,8 +84,8 @@ export default class PayloadDetailsCard extends React.PureComponent {
           )
         }
         {
-          signature && (
-            <View style={{ padding: 5, paddingVertical: 2 }}>
+          !!signature && (
+            <View style={{ padding: 5, paddingVertical: 2, alignItems: 'baseline' }}>
               <Text style={styles.label}>Signature</Text>
               <Text style={styles.secondaryText}>{signature}</Text>
             </View>
@@ -62,17 +96,89 @@ export default class PayloadDetailsCard extends React.PureComponent {
   }
 }
 
-function ExtrinsicPart({ label, style, value }) {
+function ExtrinsicPart({ label, fallback, value }) {
+  const [argNameValue, setArgNameValue] = useState();
+  const [period, setPeriod] = useState();
+  const [phase, setPhase] = useState();
+  const [sectionMethod, setSectionMethod] = useState();
+
+  useEffect(() => {
+    if (label === 'Method' && !fallback) {
+      const call = new Call(value);
+      const { args, meta, methodName, sectionName } = call;
+
+      const result = {};
+      for (let i = 0; i < meta.args.length; i ++) {
+          result[meta.args[i].name.toString()] = args[i].toString();
+      }
+
+      setArgNameValue(result);
+      setSectionMethod(`${sectionName}.${methodName}`);
+    };
+
+    if (label === 'Era' && !fallback) {
+      if (value.isMortalEra) {
+        setPeriod(value.asMortalEra.period.toString());
+        setPhase(value.asMortalEra.phase.toString());
+      }
+    }
+  }, []);
+
+  const renderEraDetails = () => {
+    if (period && phase) {
+      return (
+        <View style={{ display: 'flex', flexDirection: 'column', padding: 5 }}>
+          <View style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-around', alignItems: 'flex-end' }}>
+            <Text style={{...styles.subLabel, flex: 1}}>phase: </Text>
+            <Text style={{...styles.secondaryText, flex: 1}}>{phase}</Text>
+            <Text style={{...styles.subLabel, flex: 1}}>period: </Text>
+            <Text style={{...styles.secondaryText, flex: 1}}>{period}</Text>
+          </View>
+        </View>
+      )
+    } else {
+      return (
+        <View style={{ display: 'flex', flexDirection: 'row', flexWrap: 'wrap', padding: 5 }}>
+          <Text style={{...styles.subLabel, flex: 1}}>Immortal Era</Text>
+          <Text style={{...styles.secondaryText, flex: 3}}>{value.toString()}</Text>
+        </View>
+      )
+    }
+  }
+
+  const renderMethodDetails = () => {
+    return (
+      argNameValue && sectionMethod && (
+        <View style={{ display: 'flex', flexDirection: 'column' }}>
+          <Text style={styles.secondaryText}>
+            You are calling <Text style={styles.secondaryText}>{sectionMethod}</Text> with the following arguments:
+          </Text>
+            {
+              Object.entries(argNameValue).map(([key, value]) => { return (
+                <View key={key} style={{ display: 'flex', flexDirection: 'row', flexWrap: 'wrap', padding: 5, alignItems: 'flex-start' }}>
+                  <Text style={{...styles.subLabel, flex: 1}}>{key}: </Text>
+                  <Text style={{...styles.secondaryText, flex: 3}}>{value}</Text>
+                </View>
+              )})
+            }
+        </View>
+      )
+    );
+  }
 
   return (
-    <View style={[{ justifyContent: 'center', alignItems: 'flex-start' }, style]}>
-      <View
-        style={{ padding: 5, paddingVertical: 2 }}
-      >
+    <View style={[{ justifyContent: 'flex-start', alignItems: 'baseline' }]}>
+      <View style={{ margin: 5, padding: 5, paddingVertical: 2, width:'100%' }}>
         <Text style={styles.label}>
           {label}
         </Text>
-        <Text style={styles.secondaryText}>{value}</Text>
+        {
+          label === 'Method'
+            ? renderMethodDetails()
+            : label === 'Era'
+              ? renderEraDetails()
+              : <Text style={styles.secondaryText}>{value}</Text>
+        }
       </View>
     </View>
   );
@@ -85,17 +191,18 @@ const styles = StyleSheet.create({
     flexDirection: 'column',
     backgroundColor: colors.card_bg
   },
-  content: {},
-  footer: {
-    backgroundColor: '#977CF6',
-    flexDirection: 'row-reverse',
-    padding: 5
-  },
   label: {
     backgroundColor: colors.bg,
     color: colors.card_bg,
     textAlign: 'left', 
     fontSize: 20, 
+    fontFamily: fonts.bold,
+  },
+  subLabel: {
+    backgroundColor: null,
+    color: colors.card_bg_text,
+    textAlign: 'right', 
+    fontSize: 14, 
     fontFamily: fonts.bold,
   },
   icon: {
@@ -109,13 +216,9 @@ const styles = StyleSheet.create({
     color: colors.card_bg_text
   },
   secondaryText: {
-    textAlign: 'center',
+    textAlign: 'left',
     color: colors.card_bg_text,
     fontFamily: fonts.semiBold,
-    fontSize: 12
-  },
-  footerText: {
-    color: colors.card_bg,
-    fontFamily: fonts.bold
+    fontSize: 14
   }
 });
