@@ -18,8 +18,7 @@ export const EthereumNetworkKeys = Object.freeze({
 // genesisHash is used as Network key for Substrate networks
 export const SubstrateNetworkKeys = Object.freeze({
   KUSAMA: '0x3fd7b9eb6a00376e5be61f01abb429ffb0b104be05eaff4d458da48fcd425baf', // https://polkascan.io/pre/kusama/block/0
-  // POLKADOT: '123',
-  POLKADOT_TEST: '0xe4e7807c233645b910c8db58e99ed53dc71fbfff5bbe8a5534fb7e83db449210', // genesis hash from v0.5.1 polkadot --dev commit e086465916c2778a7dede7dc62e551d801dc12ca
+  SUBSTRATE_DEV: '0x4393a679e1830a487e8ae92733f089a80f3e24ba515b08dd8adb40fc6cedee8d', // substrate --dev commit ac6a2a783f0e1f4a814cf2add40275730cd41be1 hosted on wss://dev-node.substrate.dev .
 });
 
 const substrateNetworkBase = {
@@ -29,11 +28,11 @@ const substrateNetworkBase = {
     prefix: 2,
     title: 'Kusama'
   },
-  [SubstrateNetworkKeys.POLKADOT_TEST]: {
+  [SubstrateNetworkKeys.SUBSTRATE_DEV]: {
     color: '#ff8c00',
-    genesisHash: SubstrateNetworkKeys.POLKADOT_TEST,
+    genesisHash: SubstrateNetworkKeys.SUBSTRATE_DEV,
     prefix: 42,
-    title: 'Polkadot Dev'
+    title: 'Substrate Development'
   },
   // [SubstrateNetworkKeys.POLKADOT]: {
   //   color: '#e6007a',
