@@ -20,27 +20,25 @@ import { AsyncStorage } from 'react-native';
 import SecureStorage from 'react-native-secure-storage';
 import { accountId } from './account';
 
-const accountsStore_v1 = {
-  keychainService: 'accounts',
-  sharedPreferencesName: 'accounts'
-};
-
-export const deleteAccount_v1 = async account =>
-  SecureStorage.deleteItem(account.address, accountsStore_v1);
-
-export async function loadAccounts_v1() {
+export async function loadAccounts( version = 3 ) {
   if (!SecureStorage) {
     return Promise.resolve([]);
   }
 
-  return SecureStorage.getAllItems(accountsStore_v1).then(accounts =>
+  const accountStoreVersion = version === 1 ? 'accounts' : `accounts_v${version}`
+  const accountsStore = {
+    keychainService: accountStoreVersion,
+    sharedPreferencesName: accountStoreVersion
+  };
+
+  return SecureStorage.getAllItems(accountsStore).then(accounts =>
     Object.values(accounts).map(account => JSON.parse(account))
   );
 }
 
 const accountsStore = {
-  keychainService: 'accounts_v2',
-  sharedPreferencesName: 'accounts_v2'
+  keychainService: 'accounts_v3',
+  sharedPreferencesName: 'accounts_v3'
 };
 
 function accountTxsKey({ address, networkKey }) {
@@ -54,7 +52,7 @@ function txKey(hash) {
 export const deleteAccount = async account =>
   SecureStorage.deleteItem(accountId(account), accountsStore);
 
-export const saveAccount = account =>
+export const saveAccount = account => 
   SecureStorage.setItem(
     accountId(account),
     JSON.stringify(account, null, 0),
@@ -63,23 +61,15 @@ export const saveAccount = account =>
 
 export const saveAccounts = accounts => accounts.forEach(saveAccount);
 
-export async function loadAccounts() {
-  if (!SecureStorage) {
-    return Promise.resolve([]);
-  }
-
-  return SecureStorage.getAllItems(accountsStore).then(accounts =>
-    Object.values(accounts).map(account => JSON.parse(account))
-  );
-}
-
 export async function saveTx(tx) {
   if (!tx.sender) {
     throw new Error('Tx should contain sender to save');
   }
+
   if (!tx.recipient) {
     throw new Error('Tx should contain recipient to save');
   }
+
   await [
     storagePushValue(accountTxsKey(tx.sender), tx.hash),
     storagePushValue(accountTxsKey(tx.recipient), tx.hash),
@@ -89,11 +79,13 @@ export async function saveTx(tx) {
 
 export async function loadAccountTxHashes(account) {
   const result = await AsyncStorage.getItem(accountTxsKey(account));
+
   return result ? JSON.parse(result) : [];
 }
 
 export async function loadAccountTxs(account) {
   const hashes = await loadAccountTxHashes(account);
+
   return (await AsyncStorage.multiGet(hashes.map(txKey))).map(v => [
     v[0],
     JSON.parse(v[1])
@@ -102,6 +94,7 @@ export async function loadAccountTxs(account) {
 
 async function storagePushValue(key, value) {
   let currentVal = await AsyncStorage.getItem(key);
+
   if (currentVal === null) {
     return AsyncStorage.setItem(key, JSON.stringify([value]));
   } else {
@@ -112,10 +105,11 @@ async function storagePushValue(key, value) {
 }
 
 export async function loadToCAndPPConfirmation() {
-  const result = await AsyncStorage.getItem('ToCAndPPConfirmation');
+  const result = await AsyncStorage.getItem('ToCAndPPConfirmation_v3');
+
   return !!result;
 }
 
 export async function saveToCAndPPConfirmation() {
-  await AsyncStorage.setItem('ToCAndPPConfirmation', 'yes');
+  await AsyncStorage.setItem('ToCAndPPConfirmation_v3', 'yes');
 }
